Did a bunch of changes. Not quite done after all.
Non repeating reminders now don't delete themselves after triggering.

Repeating reminders won't be deleted anymore after clicking "stop repeating" on the notification.

Clicking "stop repeating" on the notification will now remove the notification and replace it with one that does not have the "stop repeating" option.

All reminder notifications now have an additional "delete" button which will delete the reminder from the database and remove the notification from the action bar.

Tapping a reminder notification now takes you to the detail fragment with the corresponding data filled in.

The detail fragment now has a delete button, deleting the reminder you're editing.

Replaced the confirm button in setter and detail fragment with a tick symbol button.

Adjusted the color scheme so that the color of the button text in alerts is now black.

Plus a few minor things.
Probably introduced a few new bugs too^^ I'll now in the coming days.

A few more changes are planned but I'm done for today.